[2024] Day 7.

+++ 2024/days/d07.h
@@ -4,7 +4,7 @@
 #include <fstream>
 
 namespace aoc {
-int D7P1(std::ifstream& input);
-int D7P2(std::ifstream& input);
+int64_t D7P1(std::ifstream& input);
+int64_t D7P2(std::ifstream& input);
 } // namespace aoc
 #endif  // AOC_D07
